Añadida página de login y retoques en frontend
