+++ Frontend/package.json
@@ -5,7 +5,7 @@
     "type": "module",
 
     "scripts": {
-        "start": "parcel --no-cache",
+        "start": "parcel src/index.html src/obras.html src/artistas.html --no-cache",
         "build": "parcel build"
     },
     "devDependencies": {
