+++ Frontend/src/artistas.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <script type="module" src="artistas.js"></script>
+    <script type="module" src="./artistas.js"></script>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
 </head>
 <body onload="readArtistas()">
@@ -15,7 +15,7 @@
     </nav>
 
     <nav class="nav justify-content-center">
-        <a class="nav-link active fs-3" aria-current="page" href="index.html">OBRAS DE ARTE</a>
+        <a class="nav-link active fs-3" aria-current="page" href="obras.html">OBRAS DE ARTE</a>
         <a class="nav-link fs-3" href="#">ARTISTAS</a>
     </nav>
