+++ Frontend/src/index.html
@@ -4,119 +4,45 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <script type="module" src="index.js"></script>
+    <script type="module" src="./index.js"></script>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
+
 </head>
-<body onload="readObras()">
-    <nav class="navbar bg-body-secondary">
-        <div class="container-fluid  justify-content-center">
-            <span class="navbar-brand mb-0 h1"><h1 class="display-3">GALERÍA DE ARTE</h1></span>
-        </div>
-    </nav>
+<body>
+  <nav class="navbar bg-secondary">
+    <div class="container-fluid  justify-content-center">
+      <span class="navbar-brand mb-0 h1 text-white"><h1 class="display-3">GALERÍA DE ARTE</h1></span>
+    </div>
+</nav>
 
-    <nav class="nav justify-content-center">
-        <a class="nav-link active fs-3" aria-current="page" href="#">OBRAS DE ARTE</a>
-        <a class="nav-link fs-3" href="artistas.html">ARTISTAS</a>
-    </nav>
 
-    <div class="container mt-5 shadow p3 mb-5 bg-gody rounded">
-        <table class="table table-hover">
-            <caption>Lista de obras</caption>
-            <thead>
-                <tr class="fs-5">
-                    <th>OBRA</th>
-                    <th>AUTOR</th>
-                    <th>AÑO</th>
-                    <th>ESTILO</th>
-                    <th></th>
-                </tr>
-            </thead>
-            <tbody id="tableBody">
-            </tbody>
-        </table>
-   
-        <!-- Button trigger modal -->    
-        <button type="button" id="botonCrear" class="btn btn-primary" data-bs-toggle="modal">Añadir</button>
-        <button type="button" id="botonActualizar" class="btn btn-primary" onclick="location.reload()">Actualizar</button>
-        <br><br>
-    </div>    
-
-    <!-- Modal -->
-    <div class="modal fade" id="añadirModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-       <div class="modal-dialog">
-           <div class="modal-content">
-               <div class="modal-header bg-secondary text-white">
-                   <h1 class="modal-title fs-5" id="exampleModalLabel">Añadir nueva obra</h1>
-                   <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-               </div>
-               <div class="modal-body">
-                   <form id="artistaForm" >
-                       <div class="mb-3">
-                           <label for="nombre" class="col-form-label">Nombre obra</label>
-                           <input type="text" class="form-control" id="nombre" required>
-                       </div>
-                       <div class="mb-3">
-                           <label for="autor" class="col-form-label">Autor</label>
-                           <input type="text" class="form-control" id="autor" required>
-                       </div>
-                       <div class="mb-3">
-                            <label for="año" class="col-form-label">Año</label>
-                            <input type="text" class="form-control" id="año" required>
-                        </div>
-                        <div class="mb-3">
-                            <label for="estilo" class="col-form-label">Estilo</label>
-                            <input type="text" class="form-control" id="estilo" required>
-                        </div>
-                   </form>
-               </div>
-               <div class="modal-footer">
-                   <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cerrar</button>
-                   <button type="button" class="btn btn-primary"  onclick="insertObras()">Guardar</button>
-               </div>
-           </div>
-       </div>
-   </div>    
-
-    <!--Modal actualizar obra-->
-    <div class="modal fade" id="actualizarModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header bg-secondary text-white">
-                    <h1 class="modal-title fs-5" id="exampleModalLabel">Modificar obra</h1>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+  <div class="container">
+    <div class="row justify-content-center mt-5">
+        <div class="col-md-6">
+            <div class="card">
+                <div class="card-header bg-secondary text-white">
+                    <h4 class="mb-0">Login</h4>
                 </div>
-                <div class="modal-body">
-                    <form id="artistaForm" >
+                <div class="card-body">
+                    <form>
                         <div class="mb-3">
-                            <label for="nombre" class="col-form-label">Nombre obra</label>
-                            <input type="text" class="form-control" id="nombreNuevo" >
+                            <label for="usuario" class="form-label">Usuario</label>
+                            <input type="text" class="form-control" id="usuario" placeholder="Introduce tu usuario">
                         </div>
                         <div class="mb-3">
-                            <label for="pais" class="col-form-label">Autor</label>
-                            <input type="text" class="form-control" id="autorNuevo">
+                            <label for="password" class="form-label">Contraseña</label>
+                            <input type="password" class="form-control" id="password" placeholder="Introduce tu contraseña">
                         </div>
-                        <div class="mb-3">
-                            <label for="pais" class="col-form-label">Año</label>
-                            <input type="text" class="form-control" id="añoNuevo">
-                        </div>
-                        <div class="mb-3">
-                            <label for="pais" class="col-form-label">Estilo</label>
-                            <input type="text" class="form-control" id="estiloNuevo">
-                        </div>
-                            <input type="hidden" id="idNuevo">
+                      
+                        <button type="button" class="btn btn-primary" onclick="login()">Login</button>
                     </form>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cerrar</button>
-                    <button type="button" class="btn btn-primary"  onclick="updateObra()">Guardar</button>
                 </div>
             </div>
         </div>
-    </div> 
-
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
-     
     </div>
+</div>
+    <script src="./index.js" type="module"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
 
 </body>
 </html>
